blob downloader (new); remote ais cluster
* fix remais read-range API
* dereference remais bucket in the objpost path
* add 'remais-blob-download.sh' stress test
* part nine, prev. commit: 264a54173d4c34b4ed

+++ ais/test/scripts/cp-rmnode-ec.sh
@@ -1,4 +1,10 @@
 #!/bin/bash
+
+if ! [ -x "$(command -v ais)" ]; then
+  echo "Error: ais (CLI) not installed" >&2
+  exit 1
+fi
+
 ais bucket rm ais://src-ec ais://dst -y 2>/dev/null
 
 ## create erasure-coded bucket
